third_party/WebKit/LayoutTests/platform/mac/accessibility/inline-text-bounds-for-range-expected.txt

The Hitchhiker's Guide to the Galaxy has a few things to say on the subject of resumés. A resumé, it says, is about the most massively useful thing an interstellar hitch hiker can have.

Tests that we can compute the bounds of a range of text from the accessibility tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text.length is 185
Accessible text: "The Hitchhiker's Guide to the Galaxy has a few things to say on the subject of resumés. A resumé, it says, is about the most massively useful thing an interstellar hitch hiker can have."
PASS appendedInlineText is text

Testing bounds of word: The
PASS axBounds.x is within 1 of 8
PASS axBounds.y is within 1 of 8
PASS axBounds.width is within 1 of 25
PASS axBounds.height is within 1 of 18

Testing bounds of word: Hitchhiker
PASS axBounds.x is within 1 of 37
PASS axBounds.y is within 1 of 8
PASS axBounds.width is within 1 of 68
PASS axBounds.height is within 1 of 18

Testing bounds of word: Guide
PASS axBounds.x is within 1 of 118
PASS axBounds.y is within 1 of 8
PASS axBounds.width is within 1 of 40
PASS axBounds.height is within 1 of 18

Testing bounds of word: interstellar
PASS axBounds.x is within 1 of 64
PASS axBounds.y is within 1 of 44
PASS axBounds.width is within 1 of 68
PASS axBounds.height is within 1 of 18
PASS successfullyParsed is true

TEST COMPLETE
